further refinement of bromi.mdb import procedure

+++ jacq-common/src/main/java/org/jacq/common/model/jpa/TblBotanicalObject.java
@@ -61,7 +61,7 @@ public class TblBotanicalObject implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "scientific_name_id")
-    private int scientificNameId;
+    private Long scientificNameId;
     @Column(name = "determination_date")
     @Temporal(TemporalType.DATE)
     private Date determinationDate;
@@ -136,7 +136,7 @@ public class TblBotanicalObject implements Serializable {
         this.id = id;
     }
 
-    public TblBotanicalObject(Integer id, int scientificNameId, Date recordingDate, boolean accessible, boolean redetermine, boolean separated) {
+    public TblBotanicalObject(Integer id, Long scientificNameId, Date recordingDate, boolean accessible, boolean redetermine, boolean separated) {
         this.id = id;
         this.scientificNameId = scientificNameId;
         this.recordingDate = recordingDate;
@@ -153,11 +153,11 @@ public class TblBotanicalObject implements Serializable {
         this.id = id;
     }
 
-    public int getScientificNameId() {
+    public Long getScientificNameId() {
         return scientificNameId;
     }
 
-    public void setScientificNameId(int scientificNameId) {
+    public void setScientificNameId(Long scientificNameId) {
         this.scientificNameId = scientificNameId;
     }
